% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_colors.R, R/macos_colors.R,
%   R/watchos_colors.R
\name{ios_light_pal}
\alias{ios_light_pal}
\alias{ios_dark_pal}
\alias{ios_gray_light_pal}
\alias{ios_gray_dark_pal}
\alias{macos_light_pal}
\alias{macos_dark_pal}
\alias{watchos_dark_pal}
\title{iOS Light Color Palettes}
\usage{
ios_light_pal(order = "contrast", accessible = FALSE)

ios_dark_pal(order = "contrast", accessible = FALSE)

ios_gray_light_pal(accessible = FALSE)

ios_gray_dark_pal(accessible = FALSE)

macos_light_pal(order = "contrast", accessible = FALSE, vibrant = FALSE)

macos_dark_pal(order = "contrast", accessible = FALSE, vibrant = FALSE)

watchos_dark_pal()
}
\arguments{
\item{order}{Indicates the order of colors in the palette}

\item{accessible}{Returns accessible alternative of colors}

\item{vibrant}{Returns vibrant alternative of colors}
}
\description{
Collection of color and fill scales based on watchOS color palette. watchOS uses
[a very vibrant color palette](https://developer.apple.com/design/human-interface-guidelines/watchos/visual/color/)
to highlight the information in a small screen with a black background.
}
\details{
Use this color palette with dark or pure dark themes.

![](watchOS_Dark.png "watchOS Dark Color Palette")
}
\examples{
library(scales)
scales::show_col(
    ios_light_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_light_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_dark_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_dark_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_light_pal(accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_light_pal(accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_dark_pal(accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_dark_pal(accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    macos_light_pal(order = "default", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)

scales::show_col(
    macos_light_pal(order = "contrast", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
library(scales)
scales::show_col(
    macos_dark_pal(order = "default", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)

scales::show_col(
    macos_dark_pal(order = "contrast", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
\dontrun{
library(scales)
scales::show_col(watchos_dark_pal()(9), border = NA)
}
}
