% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_inter}
\alias{font_inter}
\title{Inter font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_inter
}
\description{
`font_inter` == "`Inter`"
}
\keyword{datasets}
