% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_sf_pro}
\alias{font_sf_pro}
\title{sf_pro_font font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_sf_pro
}
\description{
`font_sf_pro` == "`SFPro`"
}
\keyword{datasets}
