% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\name{import_inter}
\alias{import_inter}
\title{Import Fonts}
\usage{
import_inter()
}
\description{
A few handy functions facilitating the process of importing and installing
custom fonts available in sfthemes.
}
\details{
sfthemes package includes 8 faces of Inter typeface: Regular, Italic, Bold,
Bold Italic, Light, Light Italic, Extra Bold, Extra Bold Italic. The
\code{import_inter} function makes sure that these fonts are
correctly installed and can be accessed by R.
}
\note{
\itemize{
\item Make sure to run this command once after installing sfthemes, otherwise,
you will encounter some warning messages regarding a missing font.
Alternatively, you can download and install your fonts manually.
}
}
