% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{import_inter}
\alias{import_inter}
\alias{font_inter}
\alias{font_inter_bold}
\alias{font_inter_light}
\title{Import Inter font for use in charts}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
import_inter()

font_inter

font_inter_bold

font_inter_light
}
\description{
\code{font_inter} == "\code{Inter}"

\code{font_inter_bold} == "\code{Inter-Bold}"

\code{font_inter_light} == "\code{Inter-Light}"
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.

\code{font_inter_bold} (a.k.a. "\code{Inter-Bold}") is not available on
Windows and will throw a warning if used in plots.

\code{font_inter_light} (a.k.a. "\code{Inter-Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
