% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios-colors.R
\name{ios_light_pal}
\alias{ios_light_pal}
\alias{ios_dark_pal}
\alias{ios_gray_light_pal}
\alias{ios_gray_dark_pal}
\title{iOS Color Palettes}
\usage{
ios_light_pal(order = "contrast", accessible = FALSE)

ios_dark_pal(order = "contrast", accessible = FALSE)

ios_gray_light_pal(accessible = FALSE)

ios_gray_dark_pal(accessible = FALSE)
}
\arguments{
\item{order}{Indicates the order of colors in the palette}

\item{accessible}{Indicates whether the accessible set of colors should be
used}
}
\description{
Set of color palettes based on
\href{https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/color/}{iOS System Colors}.
General information about each palette can be found in the
\href{../articles/colors.html}{colors} vignette.
}
\details{
These functions are ideal for accessing the raw color values, and can be
used within functions where raw color values are needed, e.g., \link{plot}.
In order to use them with \code{ggplot} functions, you must utilize one of
the followings:
\itemize{
\item \link{scale_colour_ios_light}
\item \link{scale_colour_ios_dark}
\item \link{scale_colour_ios_gray_light}
\item \link{scale_colour_ios_gray_dark}
}
}
\examples{
library(scales)
scales::show_col(
    ios_light_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_light_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_dark_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_dark_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_light_pal(accessible = FALSE)(6),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_light_pal(accessible = TRUE)(6),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_dark_pal(accessible = FALSE)(6),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_dark_pal(accessible = TRUE)(6),
    border = NA, labels = FALSE)

}
