% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_colors.R
\name{ios_gray_dark_pal}
\alias{ios_gray_dark_pal}
\title{A color palette based on iOS Light Color Palette}
\usage{
ios_gray_dark_pal(accessible = FALSE)
}
\description{
A color palette based on iOS Light Color Palette
}
\examples{
library(scales)
scales::show_col(ios_dark_pal()(9))
}
