% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_colors.R
\name{ios_light_pal}
\alias{ios_light_pal}
\title{A color palette based on iOS Light Color Palette}
\usage{
ios_light_pal(accessible = FALSE)
}
\description{
A color palette based on iOS Light Color Palette
}
\examples{
library(scales)
scales::show_col(ios_light_pal()(9))
}
