% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos_colors.R
\name{macos_light_pal}
\alias{macos_light_pal}
\alias{macos_dark_pal}
\title{macOS Color Palettes}
\usage{
macos_light_pal(order = "contrast", accessible = FALSE, vibrant = FALSE)

macos_dark_pal(order = "contrast", accessible = FALSE, vibrant = FALSE)
}
\arguments{
\item{order}{Indicates the order of colors in the palette}

\item{accessible}{Returns accessible alternative of colors}

\item{vibrant}{Returns vibrant alternative of colors}
}
\description{
Describe them
}
\examples{
library(scales)
scales::show_col(
    macos_light_pal(order = "default", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "default", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)

scales::show_col(
    macos_light_pal(order = "contrast", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_light_pal(order = "contrast", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
library(scales)
scales::show_col(
    macos_dark_pal(order = "default", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "default", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)

scales::show_col(
    macos_dark_pal(order = "contrast", accessible = FALSE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = FALSE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = TRUE)(11),
    border = NA, labels = FALSE)
scales::show_col(
    macos_dark_pal(order = "contrast", accessible = TRUE, vibrant = TRUE)(11),
    border = NA, labels = FALSE)
}
