% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos_colors.R
\name{macos_light_pal}
\alias{macos_light_pal}
\title{A color palette based on iOS Light Palette}
\usage{
macos_light_pal(accessible = FALSE, vibrant = FALSE)
}
\description{
A color palette based on iOS Light Palette
}
\examples{
\dontrun{
library(scales)
scales::show_col(macos_light_pal()(9), border = NA)
}
}
