% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios-colours.R
\name{scale_colour_ios_light}
\alias{scale_colour_ios_light}
\alias{scale_fill_ios_light}
\title{iOS Light Color Scales}
\usage{
scale_colour_ios_light(order = "contrast", accessible = FALSE, ...)

scale_fill_ios_light(order = "contrast", accessible = FALSE, ...)
}
\arguments{
\item{order}{Indicates the order of colors in the palette}

\item{accessible}{Indicates whether the accessible set of colors should be
used}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale}}
  \describe{
    \item{\code{aesthetics}}{The names of the aesthetics that this scale works with.}
    \item{\code{scale_name}}{The name of the scale that should be used for error messages
associated with this scale.}
    \item{\code{palette}}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take (e.g., \code{\link[scales:hue_pal]{scales::hue_pal()}}).}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones
}}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{na.value}}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
    \item{\code{guide}}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}
}
\description{
\strong{Note:} We recommend to not use the iOS Palette for more than 6 categories.
This guarantee an acceptable level of contrast.
}
\seealso{
Other colour scales: 
\code{\link{scale_colour_ios_dark}()},
\code{\link{scale_colour_ios_gray_dark}()},
\code{\link{scale_colour_ios_gray_light}()},
\code{\link{scale_colour_macos_dark}()},
\code{\link{scale_colour_macos_light}()},
\code{\link{scale_colour_watchos_dark}()}
}
\concept{colour scales}
