% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_colors.R
\name{scale_color_ios_dark}
\alias{scale_color_ios_dark}
\alias{scale_colour_ios_dark}
\alias{scale_fill_ios_dark}
\title{A New Thing!}
\usage{
scale_color_ios_dark(accessible = FALSE, ...)

scale_fill_ios_dark(accessible = FALSE, ...)
}
\arguments{
\item{accessible}{Returns accessible alternative of colors}
}
\description{
A New Thing!

What about this!
}
