% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos_colors.R
\name{macos_light_pal}
\alias{macos_light_pal}
\alias{macos_dark_pal}
\alias{scale_color_macos_dark}
\alias{scale_fill_macos_dark}
\title{A color palette based on iOS Light Palette}
\usage{
macos_light_pal(accessible = FALSE, vibrant = FALSE)

macos_dark_pal(accessible = FALSE, vibrant = FALSE)

scale_color_macos_dark(accessible = FALSE, vibrant = FALSE, ...)

scale_fill_macos_dark(accessible = FALSE, vibrant = FALSE, ...)
}
\arguments{
\item{accessible}{Returns accessible alternative of colors}

\item{vibrant}{Returns vibrant alternative of colors}
}
\description{
A color palette based on iOS Light Palette

A color palette based on iOS Light Palette
}
\examples{
\dontrun{
library(scales)
scales::show_col(macos_light_pal()(9), border = NA)
}
\dontrun{
library(scales)
scales::show_col(macos_light_pal()(9), border = NA)
}
}
