% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_sizes.R
\docType{data}
\name{sf_Medium_font_sizes}
\alias{sf_Medium_font_sizes}
\title{sf_Medium}
\format{
An object of class \code{list} of length 2.
}
\usage{
sf_Medium_font_sizes
}
\description{
sf_Medium
}
\keyword{datasets}
