% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{import_sf_pro}
\alias{import_sf_pro}
\alias{font_sf_pro}
\alias{font_sf_pro_bold}
\alias{font_sf_pro_light}
\title{SF Pro Font
Import SF Pro font for use in charts}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
import_sf_pro()

font_sf_pro

font_sf_pro_bold

font_sf_pro_light
}
\description{
\code{font_sf_pro} == "\code{SFPro}"

\code{font_sf_pro_bold} == "\code{SFPro-Bold}"

\code{font_sf_pro_light} == "\code{SFPro-Light}"
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.

\code{font_sf_pro_bold} (a.k.a. "\code{SFPro-Bold}") is not available on
Windows and will throw a warning if used in plots.

\code{font_sf_pro_light} (a.k.a. "\code{SFPro-Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
