% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-fonts.R
\name{sf_set_inter}
\alias{sf_set_inter}
\alias{sf_set_sf_pro}
\alias{sf_set_custom_font}
\title{Set Fonts}
\usage{
sf_set_inter(ret)

sf_set_sf_pro(ret)

sf_set_custom_font(ret, font_family_name)
}
\arguments{
\item{ret}{A \code{\link[ggplot2:theme]{ggplot2::theme()}} object}

\item{font_family_name}{The name of the font}
}
\description{
These functions can be used to change the font of all text elements in any
themes. Additionally, all font weights (or faces) of the given theme will be
set appropriately to ensure readability, e.g., setting title font
to be bold.


}
\note{
\itemize{
\item Due to limitation of \code{extrafont} and the format of SF Pro, \code{sf_set_sf_pro()}
is not yet available, and SF Pro does not work as expected.
}
}
\examples{
\dontrun{

  library(ggplot2)

  ggplot(mtcars) +
   geom_point(aes(x = wt, y = mpg, colour = factor(gear))) +
   labs(title = "Fuel Economy",
        subtitle = "Fuel economy declines as weight increases, 1973-74",
        caption = "Data from the 1974 Motor Trend US magazine.",
        x = "Weight (1000 lbs)",
        y = "MPG",
        colour = "Gears") +
   facet_grid(vs ~ am) +
   sf_set_inter(theme_linedraw()) +
   scale_colour_ios_light()

}
\dontrun{
library(ggplot2)

ggplot(mtcars) +
 geom_point(aes(x = wt, y = mpg, colour = factor(gear))) +
 labs(title = "Fuel Economy",
      subtitle = "Fuel economy declines as weight increases, 1973-74",
      caption = "Data from the 1974 Motor Trend US magazine.",
      x = "Weight (1000 lbs)",
      y = "MPG",
      colour = "Gears") +
 facet_grid(vs ~ am) +
 sf_set_custom_font(theme_linedraw(), "Gill Sans") +
 scale_colour_ios_light()
}
}
