% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_size_scales.R
\name{sf_set_text_sizes}
\alias{sf_set_text_sizes}
\title{Typography}
\usage{
sf_set_text_sizes(size = "xSmall")
}
\arguments{
\item{size}{Indicates the size class of the elements}
}
\description{
Set text sizes of different elements of a plot
}
\details{
SFthemes offers a set of scale sizes to adjust the overall
size of different elements. This includes text as well as 
graphical elements.
}
