% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_sizes.R
\docType{data}
\name{sf_xxLarge_font_sizes}
\alias{sf_xxLarge_font_sizes}
\title{sf_xxLarge}
\format{
An object of class \code{list} of length 2.
}
\usage{
sf_xxLarge_font_sizes
}
\description{
sf_xxLarge
}
\keyword{datasets}
