% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfthemes-package.R
\docType{package}
\name{sfthemes}
\alias{sfthemes}
\alias{sfthemes-package}
\title{San Francisco Themes}
\description{
An Accessible Set of Themes and Color Palettes Optimized for Light/Dark
Appearances, and Different Screen Sizes
}
\details{
sfthemes is a pair of \href{https://ggplot2.tidyverse.org}{\code{ggplot2}} themes,
\code{theme_sf_light()}, \code{theme_sf_dark()}, and a collection of color scales based
on \href{https://developer.apple.com/design/human-interface-guidelines/}{Apple's Human Interface Guidelines}.
Themes are inspired by the occasional appearances of charts used by Apple within different
contexts, e.g., Apple Health app, Screen Time on iOS and macOS. And colors are
adapted from iOS, macOS, and watchOS system colors. Each color scale has a
light and dark variant as well that compliments the light/dark themes,
\code{scale_colour_ios_light()} and \code{scale_colour_ios_dark()}, respectively. In
addition, sfthemes uses \href{https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/typography/}{dynamic type sizing}
in order to provide a consistent and elegant overall resizing and re-scaling of
plot elements. Read more \href{articles/typography.html}{here}.

When combined, sfthemes offers a unique opportunity to generate two version of
your plot with two \emph{similar} yet \emph{individually} optimized color palettes for
light and dark themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfthemes.amirmasoudabdol.name}
  \item Report bugs at \url{https://github.com/amirmasoudabdol/sfthemes/issues}
}

}
\author{
Amir Masoud Abdol (i@amirmasoudabdol.name)
}
\keyword{internal}
