% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-dark.R
\name{theme_sf_dark}
\alias{theme_sf_dark}
\title{San Francisco Dark Theme}
\usage{
theme_sf_dark(
  font_family = "Inter",
  base_size = NULL,
  plot_title_size = NULL,
  plot_title_face = NULL,
  plot_title_margin = 10,
  subtitle_size = NULL,
  subtitle_face = NULL,
  subtitle_margin = 15,
  strip_text_size = NULL,
  strip_text_face = NULL,
  caption_size = NULL,
  caption_face = NULL,
  caption_margin = 25,
  axis_text_size = NULL,
  axis_title_size = NULL,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = ios_text_on_dark_cols[["label"]],
  grid = TRUE,
  axis_col = ios_text_on_dark_cols[["label"]],
  axis = TRUE,
  ticks = TRUE,
  font_size_scale = "xSmall",
  element_size_scale = "xSmall",
  offset_x_ticks = FALSE,
  offset_y_ticks = FALSE,
  scale = NULL
)
}
\arguments{
\item{font_family}{primary font family of the plot}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{axis_text_size}{font size of axis text}

\item{axis_title_just}{axis title font justification one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid_col}{grid color}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of
`X`, `x`, `Y`, `y`)}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}

\item{font_size_scale}{Scaling fonts}

\item{element_size_scale}{Scaling element sizes}

\item{offset_x_ticks}{fancy x tick labels}

\item{offset_y_ticks}{fancy y tick labels}

\item{scale}{scale everything}
}
\description{
San Francisco Dark Theme
}
\examples{
\dontrun{
library(ggplot2); library(sfthemes)

ggplot(mpg, aes(x = displ, y = hwy, colour = class)) +
  geom_point() +
  labs(x= "Displacement", y = "Highway Miles per Gallon",
       color = "Class",
       title = "San Francisco Dark Theme",
       subtitle = "Using iOS Dark Color Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38
                     popular models of cars.") +
  theme_sf_dark() +
  scale_colour_ios_dark()

ggplot(mpg, aes(y = class)) +
  geom_bar(aes(fill = drv), color = NA) +
  labs(x= "Class", y= "Count",
       fill = "Drive",
       title = "San Francisco Dark Theme",
       subtitle = "Using Accessible iOS Dark Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38
                      popular models of cars.") +
  theme_sf_dark() +
  scale_fill_ios_dark(accessible = TRUE) +
  scale_colour_ios_dark(accessible = TRUE)

}
}
