% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchos_colors.R
\name{watchos_dark_pal}
\alias{watchos_dark_pal}
\title{watchOS Color Palette}
\usage{
watchos_dark_pal(order = "contrast")
}
\description{
Collection of color and fill scales based on watchOS color palette. watchOS uses
[a very vibrant color palette](https://developer.apple.com/design/human-interface-guidelines/watchos/visual/color/)
to highlight the information in a small screen with a black background.
}
\details{
Use this color palette with dark or pure dark themes.

![](watchOS_Dark.png "watchOS Dark Color Palette")
}
\examples{
\dontrun{
library(scales)
scales::show_col(watchos_dark_pal()(9), border = NA)
}
}
