% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchos-colours.R
\name{watchos_dark_pal}
\alias{watchos_dark_pal}
\title{watchOS Color Palette}
\usage{
watchos_dark_pal(order = "contrast")
}
\arguments{
\item{order}{Indicates the order of colors in the palette}
}
\description{
Set of color palettes based on \href{https://developer.apple.com/design/human-interface-guidelines/watchos/visual/color/}{watchOS System Colors}.
General information about each palette can be found in the \href{../articles/colors.html}{colors} vignette.
}
\details{
These functions are ideal for accessing the raw color values, and
can be used within functions where raw color values are needed, e.g.,
\link{plot}. In order to use them with \code{ggplot} functions, you must utilize one
of the followings:
\itemize{
\item \link{scale_colour_watchos_dark}
}

\strong{Note}
\itemize{
\item We recommend paining this palette with dark themes, as the high
base brightness of colors may make them hard to distinguish in light
themes
}
}
\examples{

library(scales)

 scales::show_col(watchos_dark_pal(order = "default")(10),
                  borders = NA, labels = FALSE)
 scales::show_col(watchos_dark_pal(order = "contrast")(10),
                  borders = NA, labels = FALSE)

}
