% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchos_colors.R
\docType{data}
\name{watchos_dark_palette}
\alias{watchos_dark_palette}
\title{A raw color palette based on watchOS dark color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
watchos_dark_palette
}
\description{
A raw color palette based on watchOS dark color palette
}
\keyword{datasets}
