% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_colors.R
\name{ios_light_pal}
\alias{ios_light_pal}
\alias{ios_dark_pal}
\alias{ios_gray_light_pal}
\alias{ios_gray_dark_pal}
\title{iOS Light Color Palettes}
\usage{
ios_light_pal(order = "contrast", accessible = FALSE)

ios_dark_pal(order = "contrast", accessible = FALSE)

ios_gray_light_pal(accessible = FALSE)

ios_gray_dark_pal(accessible = FALSE)
}
\arguments{
\item{order}{Indicates the order of colors in the palette}

\item{accessible}{Indicates whether the accessible set of colors should be used}
}
\description{
iOS Light Color Palettes

iOS Dark Color Palettes

iOS Grey Palette
}
\examples{
library(scales)
scales::show_col(
    ios_light_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_light_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_light_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_dark_pal(order = "default", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "default", accessible = TRUE)(9),
    border = NA, labels = FALSE)

scales::show_col(
    ios_dark_pal(order = "contrast", accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_dark_pal(order = "contrast", accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_light_pal(accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_light_pal(accessible = TRUE)(9),
    border = NA, labels = FALSE)

library(scales)
scales::show_col(
    ios_gray_dark_pal(accessible = FALSE)(9),
    border = NA, labels = FALSE)
scales::show_col(
    ios_gray_dark_pal(accessible = TRUE)(9),
    border = NA, labels = FALSE)

}
