% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_sf_pro}
\alias{import_sf_pro}
\alias{sf_set_sf_pro}
\title{San Francisco Pro Typeface}
\usage{
import_sf_pro()

sf_set_sf_pro(ret)
}
\description{
SF Pro font for use in charts
}
\details{
SF Pro Font
}
\note{
\itemize{
\item Due to limitation of \code{extrafont} and format of SF Pro, these functions
might not working as expected.
}
}
