% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_base.R, R/sf_light.R
\docType{data}
\name{font_sf_pro_text}
\alias{font_sf_pro_text}
\alias{font_sf_pro_text_light}
\alias{font_st_pro_text}
\alias{font_st_pro_text_light}
\title{SFPro font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.

length 1 character vector

An object of class \code{character} of length 1.
}
\usage{
font_sf_pro_text

font_sf_pro_text_light

font_st_pro_text

font_st_pro_text_light
}
\description{
\code{font_sf_pro_text} == "\code{SFProText}"

\code{font_sf_pro_text_light} == "\code{SFProText-Light}"

\code{font_st_pro_text} == "\code{SFProText}"

\code{font_st_pro_text_light} == "\code{SFProText-Light}"
}
\note{
\code{font_sf_pro_text_light} (a.k.a. "\code{SFProText-Light}") is not available on
Windows and will throw a warning if used in plots.

\code{font_st_pro_text_light} (a.k.a. "\code{SFProText-Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
