% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_dark.R
\name{theme_sf_dark}
\alias{theme_sf_dark}
\title{A light \link{ggplot2} theme inspired by Apple's HIG, and occasional appearance of charts in Apple's platforms, i.e.,
Apple Health, Screen Time.}
\usage{
theme_sf_dark(
  base_family = "Inter",
  base_size = 11.5,
  plot_title_family = "Inter Bold",
  plot_title_size = 18,
  plot_title_face = "plain",
  plot_title_margin = 10,
  subtitle_family = if (.Platform$OS.type == "windows") "Inter" else "Inter Light",
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = "Inter Medium",
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = if (.Platform$OS.type == "windows") "Inter" else "Inter Thin",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = 9,
  axis_title_family = base_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = ios_text_on_dark_cols[["white"]],
  grid = TRUE,
  axis_col = ios_text_on_dark_cols[["white"]],
  axis = TRUE,
  ticks = TRUE
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of \verb{[blmcrt]}}

\item{plot_margin}{plot margin (specify with \link[ggplot2:element]{ggplot2::margin})}

\item{grid_col}{grid color}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{ticks if \code{TRUE} add ticks}
}
\description{
A light \link{ggplot2} theme inspired by Apple's HIG, and occasional appearance of charts in Apple's platforms, i.e.,
Apple Health, Screen Time.
}
\examples{
library(ggplot2)
library(dplyr)

opts <- options(
  ggplot2.discrete.fill = list(scale_fill_ios_dark()),
  ggplot2.discrete.color = list(scale_color_ios_dark())
)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_sf_dark() +
  scale_color_ios_dark()

# seminal bar chart

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_sf_dark(grid="Y") +
  theme(axis.text.y=element_blank()) +
  scale_color_watchos_dark()

}
