% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-dark.R
\name{theme_sf_dark}
\alias{theme_sf_dark}
\title{San Francisco Dark Theme}
\usage{
theme_sf_dark(
  font_family = "Inter",
  plot_title_margin = 10,
  subtitle_margin = 15,
  caption_margin = 25,
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = ios_text_on_dark_cols[["label"]],
  grid = TRUE,
  axis_col = ios_text_on_dark_cols[["label"]],
  axis = TRUE,
  ticks = TRUE,
  size_class = NULL,
  font_size_class = "xSmall",
  element_size_class = "xSmall",
  offset_x_ticks = FALSE,
  offset_y_ticks = FALSE
)
}
\arguments{
\item{font_family}{Primary font family of the plot}

\item{subtitle_margin}{Plot subtitle margin bottom (single numeric value)}

\item{axis_title_just}{Axis title font justificationk one of \verb{[blmcrt]}}

\item{plot_margin}{Plot margin (specify with \link[ggplot2:element]{ggplot2::margin})}

\item{grid_col}{Grid color}

\item{grid}{Panel grid (\code{TRUE}, \code{FALSE}, or a
combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis_col}{Axis color}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{Ticks if \code{TRUE} add ticks}

\item{size_class}{Resizes texts and elements based on one of the given \href{../articles/size_classes.html}{size classes}:
\itemize{
\item A \code{character} from {"xSmall", "Small", "Medium", "Large", "xLarge", "xxLarge", "xxxLarge"}
}}

\item{font_size_class}{Only resizes text based on one of the above size classes}

\item{element_size_class}{Only resizes elements based on one of the above size classes}

\item{offset_x_ticks}{Fancy x tick labels}

\item{offset_y_ticks}{Fancy y tick labels}
}
\description{
A minimal theme with dark background inspired by — and adapted from — the
occasional appearances of charts used by Apple within different
contexts, e.g., \href{https://www.apple.com/ios/health/}{Apple Health} app,
\href{https://support.apple.com/en-us/HT208982}{Screen Time on iOS} and macOS.
}
\details{
Recommended color and fill scales to be used with:
\itemize{
\item \link{scale_colour_ios_dark}
\item \link{scale_fill_ios_dark}
\item \link{scale_colour_macos_dark}
\item \link{scale_fill_macos_dark}
\item \link{scale_colour_watchos_dark}
\item \link{scale_fill_watchos_dark}
}
}
\examples{
\dontrun{
library(ggplot2); library(sfthemes)

ggplot(mpg, aes(x = displ, y = hwy, colour = class)) +
  geom_point() +
  labs(x= "Displacement", y = "Highway Miles per Gallon",
       color = "Class",
       title = "San Francisco Dark Theme",
       subtitle = "Using iOS Dark Color Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38
                     popular models of cars.") +
  theme_sf_dark() +
  scale_colour_ios_dark()

ggplot(mpg, aes(y = class)) +
  geom_bar(aes(fill = drv), color = NA) +
  labs(x= "Class", y= "Count",
       fill = "Drive",
       title = "San Francisco Dark Theme",
       subtitle = "Using Accessible iOS Dark Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38
                      popular models of cars.") +
  theme_sf_dark() +
  scale_fill_ios_dark(accessible = TRUE) +
  scale_colour_ios_dark(accessible = TRUE)

}
}
