% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_dark.R
\name{theme_sf_dark}
\alias{theme_sf_dark}
\title{San Francisco Dark Theme}
\usage{
theme_sf_dark(
  base_family = "Inter",
  base_size = NULL,
  plot_title_family = "Inter Bold",
  plot_title_size = NULL,
  plot_title_face = "plain",
  plot_title_margin = 10,
  subtitle_family = if (.Platform$OS.type == "windows") "Inter" else "Inter Light",
  subtitle_size = NULL,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = "Inter Medium",
  strip_text_size = NULL,
  strip_text_face = "plain",
  caption_family = if (.Platform$OS.type == "windows") "Inter" else "Inter Thin",
  caption_size = NULL,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = NULL,
  axis_title_family = base_family,
  axis_title_size = NULL,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = ios_text_on_dark_cols[["white"]],
  grid = TRUE,
  axis_col = ios_text_on_dark_cols[["white"]],
  axis = TRUE,
  ticks = TRUE,
  font_size_scale = "xSmall",
  element_size_scale = "xSmall",
  offset_x_ticks = FALSE,
  offset_y_ticks = FALSE,
  scale = NULL
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid_col}{grid color}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}

\item{font_size_scale}{Scaling fonts}

\item{element_size_scale}{Scaling element sizes}

\item{offset_x_ticks}{fancy y tick labels}

\item{scale}{scale everything}
}
\description{
San Francisco Dark Theme
}
\examples{
library(tidyverse)

# seminal scatterplot
 ggplot(mpg, aes(x = displ, y = hwy, colour = class)) +
  geom_point() +
  labs(x="Displacement", y="Highway Miles per Gallon",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_sf_dark() +
  scale_color_ios_dark()

# seminal bar chart

ggplot(mpg, aes(class)) +
  geom_bar(aes(fill = drv)) +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_sf_dark(grid="Y") +
  theme(axis.text.y=element_blank()) +
  scale_color_ios_dark(accessible = TRUE)

}
