% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_light.R
\name{theme_sf_light}
\alias{theme_sf_light}
\title{San Francisco Light Theme}
\usage{
theme_sf_light(
  base_family = "Inter",
  base_size = NULL,
  plot_title_family = "Inter Bold",
  plot_title_size = NULL,
  plot_title_face = "plain",
  plot_title_margin = 10,
  subtitle_family = "Inter",
  subtitle_size = NULL,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = "Inter Medium",
  strip_text_size = NULL,
  strip_text_face = "plain",
  caption_family = "Inter",
  caption_size = NULL,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = NULL,
  axis_title_family = base_family,
  axis_title_size = NULL,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = TRUE,
  ticks = TRUE,
  font_size_scale = "xSmall",
  element_size_scale = "xSmall",
  offset_x_ticks = FALSE,
  offset_y_ticks = FALSE,
  scale = NULL
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid_col}{grid color}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}

\item{font_size_scale}{Scaling fonts}

\item{element_size_scale}{Scaling element sizes}

\item{offset_x_ticks}{fancy x tick labels}

\item{offset_y_ticks}{fancy y tick labels}

\item{scale}{scale everything}
}
\description{
San Francisco Light Theme
}
\examples{
library(ggplot2)

ggplot(mpg, aes(x = displ, y = hwy, colour = class)) +
  geom_point() +
  labs(x="Displacement", y="Highway Miles per Gallon",
       color = "Class",
       title="San Francisco Light Theme",
       subtitle="Using iOS Light Color Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38 popular models of cars.") +
  theme_sf_light() +
  scale_color_ios_light()

ggplot(mpg, aes(class)) +
  geom_bar(aes(fill = drv, color = NA)) +
  labs(x="Class", y="Count",
       fill = "Drive",
       title="San Francisco Light Theme",
       subtitle="Using Accessible iOS Light Palette",
       caption = "Fuel economy data from 1999 to 2008 for 38 popular models of cars.") +
  theme_sf_light() +
  scale_fill_ios_light(accessible = TRUE) +
  scale_color_ios_light(accessible = TRUE)

}
\concept{themes sflight}
