% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchos_colors.R
\name{watchos_dark_pal}
\alias{watchos_dark_pal}
\title{watchOS Color Palette}
\usage{
watchos_dark_pal(order = "contrast")
}
\arguments{
\item{order}{Indicates the order of colors in the palette}
}
\description{
watchOS Color Palette
}
\examples{
library(scales)
scales::show_col(
    watchos_dark_pal(order = "default")(9),
    border = NA, labels = FALSE)
scales::show_col(
    watchos_dark_pal(order = "contrast")(9),
    border = NA, labels = FALSE)
}
